Test plugin side effects rather than jobs
